+++ lapce-proxy/src/plugin/catalog.rs
@@ -324,6 +324,9 @@ impl PluginCatalog {
                 }
                 self.check_unactivated_volts();
             }
+            UpdatePluginConfigs(configs) => {
+                self.plugin_configurations = configs;
+            }
             PluginServerLoaded(plugin) => {
                 // TODO: check if the server has did open registered
                 if let Ok(ProxyResponse::GetOpenFilesContentResponse { items }) =
@@ -355,6 +358,17 @@ impl PluginCatalog {
                     let _ =
                         install_volt(catalog_rpc, workspace, configurations, volt);
                 });
+            }
+            ReloadVolt(volt) => {
+                let volt_id = volt.id();
+                let ids: Vec<PluginId> = self.plugins.keys().cloned().collect();
+                for id in ids {
+                    if self.plugins.get(&id).unwrap().volt_id == volt_id {
+                        let plugin = self.plugins.remove(&id).unwrap();
+                        plugin.shutdown();
+                    }
+                }
+                let _ = self.plugin_rpc.unactivated_volts(vec![volt]);
             }
             StopVolt(volt) => {
                 let volt_id = volt.id();